requirements.txt
# Core
numpy
pandas
tqdm
python-dotenv

# OCR / PDF
pytesseract
pdfplumber
pillow

# Embeddings / Vector DB
sentence-transformers
faiss-cpu

# LLM / Local inference
requests

# Utilities
pyyaml
